events: the stream is cards:done, and an absent cost stays absent (#2563, Johnny HOLD 4/10)
events.Stream is cards:done, the stream record and ci already read; the event
fields are added to its entries, and the SQLite file is a fold of that one
stream, rebuildable from it, not a second record. An entry written before the
fields were added (no `event`) is counted as skipped, never guessed into ok/fail.

tokens_in, tokens_out and usd are *int64/*float64 on Event: nil writes no field,
FromFields leaves an absent field nil, the fold stores NULL (columns nullable, no
default), dump/report print a dash, and totals no longer COALESCEs usd to 0.
`nova-pulse event` writes a cost only when its flag was given. The test that
asserted zero is replaced by TestMissingCostStaysAbsentAndBadNumbersAreRefused,
plus TestTheStreamIsCardsDone, TestAMissingCostFoldsToNullNotZero and
TestEventWithoutACostWritesNoCostField.

+++ internal/events/redis_integration_test.go
@@ -18,7 +18,7 @@ import (
 // built with `nova-secrets exec --only NOVA_REDIS_BENCH_PASSWORD`, so no credential is ever
 // in this tree and no unit test ever dials anything (AGENTS.md rule 2).
 //
-// Every key it writes is under `ev:test-<stamp>`, never `ev:cards`, so a live sprint's
+// Every key it writes is under `ev:test-<stamp>`, never `cards:done`, so a live sprint's
 // stream cannot be touched by a test run, and the stream is deleted at the end.
 const (
 	testEnable   = "NOVA_REDIS_TEST"
@@ -101,9 +101,9 @@ func TestAHundredDonesAgainstTheFleetStore(t *testing.T) {
 			Model:     []string{"fable", "sonnet"}[i%2],
 			Route:     []string{"studio", "hulk"}[i%2],
 			Kind:      OK,
-			TokensIn:  1000,
-			TokensOut: 100,
-			USD:       0.01,
+			TokensIn:  Int64(1000),
+			TokensOut: Int64(100),
+			USD:       Float64(0.01),
 		}); err != nil {
 			t.Fatalf("emitting DONE %d: %v", i, err)
 		}
